pipeline/requirements.txt
python-dotenv
sqlalchemy
pymysql
boto3
pandas
pylint
